モデルテストコード

+++ db/migrate/20210812062242_create_attendances.rb
@@ -4,7 +4,7 @@ class CreateAttendances < ActiveRecord::Migration[6.0]
 
       t.references :user,          null: false, foreign_key: true
       t.integer :arrive_id,      null: false
-      t.integer :flying_id
+      t.float :flying_id
       t.date :attendance_day,  null: false
 
       t.timestamps
